更新应用版本号至 2.5.3-custom-bate，调整键码表大小并新增自定义键码表支持

+++ src/device.h
@@ -56,7 +56,8 @@ struct device_event {
 		DEV_REMOVED,
 	} type;
 
-	uint8_t code;
+	// uint8_t code;
+	uint16_t code;
 	uint8_t pressed;
 
 	int32_t x;
